Stop a KPI reconfiguration from carrying stale totals, and unbrick the endTime read
Two findings from the `1875ba6..ac7eb44` review.

`setKpiConfig` rewrote the whole `KpiConfig` but left `lastScannedBlock` and every
`verifiedTotals` entry untouched. Correcting a KPI configured against the wrong contract
therefore resumed the relayer at `N+1`, so `[windowStartBlock, N]` was never rescanned for
the right event — while the totals folded from the wrong one stayed live as the ceiling a
claim is trimmed to. Raising `windowStartBlock` left the same gap in reverse, and changing
`scale` re-denominated every stored total retroactively.

`KpiConfig` now carries an `epoch` that is part of every total's storage key, bumped
whenever the watched event changes and left alone when only `windowEndBlock` moves — the
one case replacement exists for. A bump abandons every figure observed under the old config
in a single write and resets the checkpoint. Clearing the entries instead is not available:
`verifiedTotals` is a mapping with no enumerable key set, so a user who only ever appeared
under the wrong config would never be overwritten at all.

`AttributionRegistry._requireCampaignOpen` decoded `endTime()` as `uint64`, which reverts on
a non-zero upper word. The `status()` read four lines below was deliberately decoded as
`uint256` for exactly this reason and says so; the reasoning had been applied to one read and
not its neighbour. A registrant returning a full 32-byte value could therefore brick every
`storeTouch` naming it, contradicting the documented reading that a campaign answering
neither call is simply unbounded. Both that read and `attributionWindow` in
`_effectiveMaxDuration` now decode as `uint256` and compare numerically, and `CampaignOver`
carries the full word so it cannot report a garbage endTime as a plausible one.

414 forge tests pass. The verifier's contract-level and struct docs also carry an in-flight
comment trim from a concurrent session in this working tree.

+++ src/interfaces/IAttributionRegistry.sol
@@ -22,7 +22,11 @@ interface IAttributionRegistry {
     error InvalidSignature();
     error PromoterNotRegistered(address campaign, bytes32 promoterId);
     error ZeroWindow();
-    error CampaignOver(uint64 endTime, uint64 timestamp);
+    /// @dev `endTime` is the full 32-byte word the campaign answered with, not a `uint64`, because the
+    ///      registry reads it from an untyped staticcall against a registrant that need not be a
+    ///      `Campaign`. Narrowing it here would either revert on a dirty word or report a garbage
+    ///      endTime as a plausible one. Same reasoning as `CampaignTerminal`.
+    error CampaignOver(uint256 endTime, uint64 timestamp);
     error CampaignTerminal(uint256 status);
 
     // ── events ───────────────────────────────────────────────────
